Move youtube command to mediaplayer
